eod changes

+++ application/models/client/client_user_model.php
@@ -14,8 +14,13 @@ class Client_User_Model extends CI_Model {
     public $password = NULL;
 
     function create_user($data) {
-
-
+            $user_following = json_encode(array());
+             if ((session('profile_pic'))) {
+            $profile_pic = json_encode(array(session('profile_pic')));
+        } else {
+            $profile_pic = "";
+        }
+        
         $table_array = array(
             'name' => $data['full_name'],
             'email' => $data['email'],
@@ -26,7 +31,11 @@ class Client_User_Model extends CI_Model {
             'interested_domain_id' => $data['domain'],
             'country_id' => $data['country'],
             'preferred_destination' => $data['preferred_destination'],
+            'profile_pic' => $profile_pic,
             'created_date' => get_date(),
+            'user_following' => $user_following,
+            'gender' => session('gender'),
+            'company' => session('company'),
         );
 
         if ($data['profile_type'] == 'normal') {
@@ -37,8 +46,10 @@ class Client_User_Model extends CI_Model {
             
         }
         $this->db->insert('user', $table_array);
-
-        return $this->db->insert_id();
+        $max_id = $this->db->insert_id();
+        // insert_new_user_notification($data['full_name'],$max_id);
+        return $max_id;
+       
     }
 
     function getUserData() {
@@ -50,35 +61,12 @@ class Client_User_Model extends CI_Model {
             $this->db->where('U_T.password', md5($this->password));
         }
 
+        if($this->email)
+        {
+            $this->db->where('U_T.email', $this->email);
+        }
         $query = $this->db
-                ->select('
-																							U_T.id,
-																							U_T.name,
-																							U_T.dob,
-																							U_T.gender,
-																							U_T.profile_pic,
-																							U_T.email,
-																							U_T.password,
-																							U_T.status,
-																							U_T.user_type,
-																							U_T.job_position_id,
-																							U_T.company,
-																							U_T.functionality_id,
-																							U_T.interested_degree_id,
-																							U_T.interested_field_id,
-																							U_T.interested_domain_id,
-																							U_T.industry_id,
-																							U_T.country_id,
-																							U_T.preferred_destination,
-																							U_T.facebook,
-																							U_T.linkedin,
-																							U_T.twitter,
-																							U_T.google_plus,
-																							U_T.password_key,
-																							U_T.unique_access_key,
-																							
-																				
-																				')
+                ->select('*')
                 ->from('user as U_T')
                 //-> where('U_T.id',$user_id)
                 ->where('U_T.status', 1)
@@ -124,38 +112,60 @@ class Client_User_Model extends CI_Model {
     }
 
     function save_user($data) {
-        //display($data);
+        
 
         if (isset($data['profile_pic'])) {
-
-            $profile_pic = json_encode($data['profile_pic']);
+          
+           
+            $profile_pic = json_encode(($data['profile_pic']));
+           
+             
         } else {
             $profile_pic = "";
         }
 
         $user_array = array(
             'name' => $data['name'],
-            'dob' => $data['dob']!="" ? date('Y-m-d', strtotime($data['dob'])) :"",
+            'dob' => @$data['dob']!="" ? date('Y-m-d', strtotime($data['dob'])) :"",
             'gender' => $data['gender'],
             'profile_pic' => $profile_pic,
             'country_id' => $data['country_id'],
             'preferred_destination' => $data['preferred_destination'],
-            'job_position_id' => $data['job_position_id'],
+            'job_position_id' => @$data['job_position_id'],
             'company' => $data['company'],
             'industry_id' => $data['industry_id'],
-            'functionality_id' => $data['functionality_id'],
+            'functionality_id' => @$data['functionality_id'],
             'interested_degree_id' => $data['interested_degree_id'],
             'interested_field_id' => $data['interested_field_id'],
             'interested_domain_id' => $data['interested_domain_id'],
             'modified_date' => get_date()
         );
+        
+        if(isset($data['social_type']))
+        {
+            if($data['social_type'] == 'facebook')
+            {
+                $user_array['facebook'] = $data['public_profile_url'];
+            }
+            elseif($data['social_type'] == 'linkedin')
+            {
+                $user_array['linkedin'] = $data['public_profile_url'];
+            }
+            elseif($data['social_type'] == 'google_plus')
+            {
+                $user_array['google_plus'] = $data['public_profile_url'];
+            }
+            
+        }
+        
+        //display($user_array);die;
         $this->db->where('id', $this->user_id);
         $this->db->update('user', $user_array);
-
-
-        $this->db->delete('user_education_info', array('student_id' => $this->user_id));
-
+        insert_new_user_notification($data['name'],$this->user_id);
+        //      show_query();
         if (isset($data['schoolname']) && ($data['schoolname'] != 0) && (($data['schoolname']) != "")) {
+            
+            $this->db->delete('user_education_info', array('student_id' => $this->user_id));
 
             $count_name = count($data['schoolname']);
 
@@ -190,20 +200,23 @@ class Client_User_Model extends CI_Model {
                 }
             }
         }
-        $this->db->delete('user_test_info', array('user_id' => $this->user_id));
-        $count_test_name = count(($data['tsetscore']));
-        //echo $count_test_name;
-        for ($j = 0; $j <= $count_test_name - 1; $j++) {
-            //echo $j;
-			
-            if (($data['test_name'][$j] != "") && ($data['tsetscore'][$j] > 0)) {
-                $test_info = array
-                    (
-                    'user_id' => $this->user_id,
-                    'test_name' => $data['test_name'][$j],
-                    'test_score' => $data['tsetscore'][$j]
-                );
-                (($data['test_name'][$j] != "") && ($data['tsetscore'][$j] != "" )) ? $this->db->insert('user_test_info', $test_info) : '';
+        
+        if(isset($data['tsetscore'])){
+            $this->db->delete('user_test_info', array('user_id' => $this->user_id));
+            $count_test_name = count(($data['tsetscore']));
+            //echo $count_test_name;
+            for ($j = 0; $j <= $count_test_name - 1; $j++) {
+                //echo $j;
+
+                if (($data['test_name'][$j] != "") && ($data['tsetscore'][$j] > 0)) {
+                    $test_info = array
+                        (
+                        'user_id' => $this->user_id,
+                        'test_name' => $data['test_name'][$j],
+                        'test_score' => $data['tsetscore'][$j]
+                    );
+                    (($data['test_name'][$j] != "") && ($data['tsetscore'][$j] != "" )) ? $this->db->insert('user_test_info', $test_info) : '';
+                }
             }
         }
     }
@@ -311,7 +324,117 @@ class Client_User_Model extends CI_Model {
             $this->db->delete('college_review_rating');
             return true;
         }
+       
     }
+    
+    function get_industry($industry = '')
+    {
+        if($industry)
+            $this->db->where("name LIKE '%".$industry."%'"); 
+        $query = $this->db->select('*')
+                 -> from('industry')->get();
+        $result = $query->row();
+        return $result;
+                
+    } 
+    
+    
+    
+     function user_following_details($id)
+    {
+         
+         $query = $this->db->select('user_following')
+                 -> from('user')
+                 ->where('id',$id)
+                 ->get();
+        return $query->result_array();
+          
+    } 
+    
+    
+     function user_community($id)
+    {
+         
+         $query = $this->db->select('members')
+                 -> from('communities')
+                 ->where('id',$id)
+                 ->get();
+        return $query->result_array();
+          
+    } 
+    
+     function user_community_details($id,$user_communi)
+    {
+         
+        $data = array(
+               'members' => $user_communi,
+            
+            );
+           
+            $this->db->where('id', $id);
+            $this->db->update('communities', $data); 
+    } 
+    
+    
+     function save_following_details($id,$user_following)
+    {
+         
+        $data = array(
+               'user_following' => $user_following,
+            
+            );
+           
+            $this->db->where('id', session('client_user_id'));
+            $this->db->update('user', $data); 
+          
+    } 
+    
+    
+    function delete_user_save_school($id)
+    {
+        
+        $this->db->where('id', $id);
+        $this->db->delete('user_interested_in_school'); 
+    }
+    
+    
+     function select_user_community_detail($id)
+    {
+        
+        $query = $this->db->select('id,cname')
+                 -> from('communities')
+                 ->like('members',$id)
+                 ->get();
+        return $query->result_array();
+          
+    }
+    
+    
+      function select_user_following_u($id)
+    {
+        
+        $query = $this->db->select('id,name')
+                 -> from('user')
+                 ->like('user_following',$id)
+                 ->get();
+        //show_query();
+        return $query->result_array();
+          
+    }
+    
+    
+    function user_followu($id)
+    
+    
+    {
+         
+         $query = $this->db->select('user_following')
+                 -> from('user')
+                 ->where('id',$id)
+                 ->get();
+        return $query->result_array();
+          
+    } 
 
 }
